Rework Wisdom penalty for throwing weapons improperly
A common player behavior is to throw away unwanted items, sometimes out
of convenience, sometimes to save turns, sometimes because it's just
more cathartic to throw things instead of walking a few squares and
dropping them.

When the item happened to be a nonmissile weapon or ammo without the
proper launcher, there was a 20% chance of abusing Wisdom even if the
player was doing an obvious throwaway. This commit adds in a check for
when it is NOT an obvious throwaway - if there is a monster nearby along
the intended flight path, or a invisible perceived monster marker. When
that criteria isn't met, you still get informed that this is a clumsy
throw, but Wisdom is never abused. To compensate, when you DO appear to
be trying to attack something, Wisdom is now always abused.

I'm still on the fence about whether the check for a monster in line of
the throw should get applied as a criterion for prompting the paranoid
throw prompt, so on an "obvious throwaway" you don't have to trip over
messages and prompts about it. My eventual decision was that you should
still be told that this isn't a good way to throw weapons.