Open up center of "triple rhombus" themed room
I had this one line change sitting around for a while and forgot what
the original impetus was for making it, but if I remember right, it's
because the room tried to generate as a shop and the diagonal choke
point meant the shop couldn't be completely filled properly. Rather than
making it three separate regions, open it up so the entire floor area is
orthogonally connected.