Block most teleport and levelport within Quest while nemesis lives
The impetus here is to prevent the player from using 2 controlled
levelports and 1 controlled teleport to zip straight to the nemesis,
kill them, get the Bell and quest artifact, and leave. It's not an
incredibly common strategy, but it lets you avoid a lot of potential
peril and exploration on the intervening levels. Where speedrunning is
concerned, I hope this will make its strategy a bit more interesting
than that (speedrun records between xnethack and vanilla aren't
inherently comparable, either).

Note that on a number of filler levels (mostly in the quests which
haven't yet been redesigned) the floors are still diggable. That will
likely change as I redesign them, because that usually turns them into a
horizontal traverse rather than a downward descent. I believe there
aren't any locate or home levels that lack a hardfloor flag, but didn't
check exhaustively.

Note also that upwards levelporting in the Quest, as an escape, is still
possible. Random levelports still try to go up or down (rather than
picking a level that's always higher) so they're not reliable unless
you're on the goal level (and even then you might shudder due to
levelporting to the same level).

Closes #123