Fix: panic from dropped potion of monster detection vapors
This had a simple enough cause - since I added peffects() calls into a
couple of potionbreathe() cases, it opened up the possibility of calling
strange_feeling() (which for some reason, uses up the item it's called
with) on an item that was neither thrown nor kicked, but had already
been removed from object lists waiting to be deleted, so the useup()
panicked.

There was not a clear and obvious solution. I considered making
strange_feeling not use up the object so that its callers were
responsible for doing it, but figured it would lead to annoying merge
conflicts down the line. The proposed fix of setting thrownobj to the
dropped potion in drop() so strange_feeling() wouldn't use it up was
closer, but didn't cover all cases (i.e. wishing for a potion with full
inventory while levitating and dropping it). I ended up taking this
approach, but setting thrownobj in potionbreathe() - thus covering any
other potion that might call peffects() into strange_feeling() at some
point in the future.

This bug also existed for potions of object detection, but probably
never happened in practice since it's rare to have no objects on the
level to detect. The fix covers that too.

Fixes #205