Several enhancements to the Valkyrie locate level
First, complete a forgotten TODO intended for the 8.0 release that there
should be a pager message when finding the Bifrost bridge restored. I
added this as a new qt_pager message that currently only is present in
the Valkyrie quest text. It's a bit clunky to kill Surtur on the
locate level and get his death pager immediately followed by this new
one, but that should rarely happen.

Second, there was some feedback that Hrymr's +8 sword strikes
unexpectedly hard. I don't think this needs any mechanical adjusting,
but I decided to signal that it's not an ordinary sword by naming it
Ice, which will get shown in messages. (Above and beyond it being a
platinum sword.) The name "Ice" is a reference to Eddard Stark's
Valyrian greatsword in A Song of Ice and Fire - this is similarly a
special but not actually artifact-level large sword.

Third, prevent the locate level from making bones... if it has been
visited after Surtur is dead. Basically, if the bridge is restored, it's
not suitable to leave bones to someone whose game shouldn't have it
restored. I implemented this so that ANY level which has
visited_after_event set will be bones-ineligible, but currently only the
Valkyrie locate level sets this flag.

While testing this, I noticed that if one defeats Surtur before ever
loading the locate level, the bridge doesn't properly get restored, but
this bug should be latent since I don't know of any way to bypass the
locate level now that non-wizmode levelport won't work.

Unrelated: also move the changelog entries for the recent giant changes,
which were incorrectly placed under interface rather than gameplay
changes, to the latter section.