Reduce the effect of luck on to-hit
This is a slightly modified change made by many other variants to
address the problem of maxed-out Luck completely wiping out any other
factors in most mid to late game to-hit calculations. EvilHack,
SporkHack, and NerfHack (and possibly others I don't know about) all
implement a change where Luck/3 is added instead of Luck. SpliceHack
removed Luck from the equation completely, and that was the initial plan
here, but I feel better going with something more widely tested.

However, there is that modification. What I don't like about Luck/3 is
that it mutes the influence of early game Luck-adjusting things such as
the full moon, killing a pet, converting an altar or failing, and so on.
What I did instead was keep the scaling factor the same while shifting
its ranges down so that +1 or -1 Luck is still a +1 or -1 to hit. (And,
notably, if your luck is timing out to -4 because your quest leader has
died, that will be a -2 to hit.)