High-skill martial arts critical hits disarm the opponent
This is one of several buffs to martial arts that probably won't have
outsized impact on combat outcomes, but will probably make the playstyle
more fun.

Like vampire staking, this is subject to a litany of common-sense
checks, like the weapon not being cursed and welded. The disarm uses a
weaker version of the scattering logic I employed for Yeenoghu's swats,
randomly scattering the weapon 1 square away. I used MAY_HITMON instead
of MAY_HIT because I figured the monk is skilled enough to dodge the
loose weapon if it flies towards them.