Grant AC bonus for high unarmed combat skill and no body armor/shield
This is intended as an added incentive for martial arts users to forego
armor, since if they manage to overcome the cumbersome armor to-hit
penalty, there isn't really any downside. In particular, a monk who
levels martial arts up to Grand Master skill will enjoy a +5 AC bonus.

Since it's also granted to bare-handed combat users, it's also a slight
compensation if such a hero decides not to wear armor (though there are
probably few in such a position - weaponless nudists are a bit easier
now, I guess).

Initially when I implemented this, martial arts doubled the skill bonus
rather than adding 1 to it, so Grand Master monks would get +8 AC. That
seemed too high to start with, since Grand Master can be achieved in the
mid-game if not earlier. If this change doesn't appreciably make a
difference in monk games, I may revisit the doubling.

I thought EvilHack had this feature, but instead it has a bonus based on
Dexterity. One aspect of that implementation that makes some sense, but
which I didn't do here, is to suppress the AC bonus when encumbered at
all. This is mostly for performance; I didn't want to add another
inventory loop on every player action.