Fix: iron doors exploding in flames weren't destroyed
I believe this is a consequence of doortrapped() being older than iron
doors, and not accounting for them - it assumed the fiery explosion
would destroy the door so it didn't have to, but that of course isn't
true of iron doors. Force the doorstate to be D_NODOOR afterwards
(should have no effect for wood doors).

While here, I also tripled the damage for iron door explosions compared
to wooden door explosions, since flying shards of metal are probably
worse to be hit with than splinters.