Provide artifact gen_spe and gift_value for xNetHack artifacts
Merge followup 2. These were left zero in the merge so I could elaborate
on the rationale for the values in their own commit.

Mirror Brand generates as +1 randomly or in gifts, the only Brand to do
so, but this is because it's not super effective versus as big a set of
foes as the other two are. It requires a sacrifice of difficulty 5 to
obtain, the same as Fire Brand. 5 seems a bit low, but I don't think it
should be higher than Fire Brand.

The Apple of Discord requires a sacrifice of difficulty 7 to obtain.
Slotless hungerless conflict is pretty powerful. (gen_spe is still 0
because that is meaningless for an apple).

The Amulet of Storms requires a sacrifice of only difficulty 2 to
obtain, since it's not that powerful. (Again, gen_spe is 0 because
that's meaningless for an amulet.)