Beef up checks against dragon_armor_handling being called incorrectly
Merge followup 3. The 3.7 additional dragon armor extrinsics weren't
new in this past merge, but I flagged dragon_armor_handling as not
sufficiently checking against things that might mistakenly be attempted
in xNetHack, like being called on dragon scales in the cloak slot.

Enforce the assumption that it will not be called with a null pointer,
or a non-suit piece of armor, by giving impossibles in those cases.