Get rid of magic numbers in doortrapped()
Dating back to the initial implementation of this function, there were
some not-great uses of its accepting 0,1,2 as a parameter and returning
a (semantically different) 0,1,2 as output. For instance, using
!doortrapped(...) to mean "door state didn't change", or
doortrapped(...) < 2 to mean "door wasn't destroyed by a trap", or
doortrapped(_, _, _, _, _, 2) to mean "do both pre and post door traps".

Fix the return value problem by returning values from an enum, and
explicitly comparing the result of the doortrapped() call to those
values.

Fix the parameter problem by making pre/post flags for that parameter,
and defining an additional alldoortrapped() convenience macro that means
"do both pre-action and post-action traps".