Use pline_xy and pline_mon in doortrapped()
I intend to start migrating plines added in xNetHack to pline_xy or
pline_mon for accessibility, but don't have a plan in place to do that
comprehensively. This commit only does the ones in doortrapped().

pline_mon doesn't appear to handle cases where the monster is youmonst,
so I had to use pline_xy in a couple cases.