Fix: deepest level calculation didn't factor in Gehennom overhaul
When I shortened Gehennom in xNetHack 8.0, I did not update this code
that's meant to remove the reliance of minscore games on lucking into
the shallowest possible dungeon.

Ideally we would compute the minimum dungeon depth, but there's not
really a great way to do that because the parameters for the dungeon
don't persist throughout the game and adding a savefile field to store
it after computing it at the beginning of the game seems like overkill.
So it remains hardcoded for now.