Provide obj material as string in obj.totable()
This started out as a fix for the hardcoded GLASS=19 case in a themed
room, which I mistakenly thought was xNetHack-specific, but is actually
vanilla. So that turned into pull request #1375 for vanilla, and
assuming it gets merged, I will pick up the fix at some later merge into
xNetHack.

The only thing that remains to be done here is to allow getting the
material of a specific object should the Lua designer want to.

Note: the themed room that checks the object class material is fine
obtaining the objclass's material and not the specific object's
material, because the objects it's checking for glass can only be
scrolls, rings, and wands, which aren't subject to receive different
materials than their objclass's default.