Fix: player credited for door trap kills from monsters' wand zaps
More doortrapped() changes. This resolves a longstanding bug noted in
a comment: in an admittedly very unlikely case where a monster zaps a
wand, it triggers a door trap, and the door trap kills a monster, the
player might be blamed for it, because doortrapped() treated a monst
parameter of null as the player doing it. It was possibly even limited
to certain types of door trap, since the explosion one ought to credit
deaths based on context.mon_moving already.

This is based in certain zap code not passing around a pointer to the
monster who initiated the zap; by the time zap_over_floor is being
called you're at least a few calls deep. What I did to resolve the
problem was to change doortrapped() so that a null monster means we
don't know and credit the player if it's the player's turn.