Restore hero's ability to sacrifice when hallucinating
Merge followup 5. In vanilla, sacrificing was blocked for confusion and
stunning as well as hallucinating. I don't think that change was well
thought out because the game has a bunch of YAFMs that can only be
obtained by sacrificing while hallucinating, but more importantly for
xNetHack, permahallu conduct exists and it would be impossible to win
the game if you couldn't sacrifice the Amulet.

This keeps the inability to sacrifice when confused or stunned, which I
don't have a big problem with.