Switch condititions around for undiscovered bones level sound
Traversing the cemetery chain in the current level's mapseen once per
turn is preferable to generating a random number once per turn when most
of the time it will not even be used. Same reason as why other
dosounds() conditions put the rn2 call at the very end of the
conditional.