Fix: shapechanger in seducer form would introduce itself
Applied to both nymphs and foocubi. I feel like a shapechanger which
normally has no name isn't going to suddenly pick a name for itself when
it changes into a form that can seduce, and it would also be jarring to
be attacked by an "owlbear" with a non-player-bestowed nymph or demon
name it got by introducing itself in a previous shape.

Also, if they don't introduce themselves, a savvy player can notice and
conclude they're dealing with a shapechanger.