Give bone items a +1 damage bonus against incorporeal monsters
The impetus for this change is to fix a bug: barehanded attacks on
ghosts or shades while wearing a bone (ivory) ring or bone gloves passed
harmlessly through them, which breaks the rule that if there's something
that can touch them, like silver, the attack gets to land normally and
deal its normal damage.

This is not the ideal way to fix this bug. Ideally, bone items would not
adjust damage but merely allow you to hit a ghost or shade for the normal
amount of damage, which is what I initially tried implementing: a fifth
argument to special_dmgval() that indicates whether the attack should
not pass harmlessly through incorporeal monsters, even though the return
value is 0. Sadly, this didn't work, because other functions get called
that would lack that context, and which negate damage and produce a
harmless miss message if there is no special damage.

The damage bonus fixes all these problems by ensuring there is special
damage, isn't too bad because it only works against two monsters, and
potentially opens up some more avenues of strategy against those two
monsters. So it will work for now.