Swap vrocks to green color and erinyes to orange
Merge followup 8. Vanilla changed the vrock tile to be greener and made
vrocks green as an indication of how they might leave a puff of poison
gas behind when fleeing. This makes that color consistent with vanilla.

Erinyes were also green in xNetHack, so to avoid the color conflict move
them to orange, which is now vacant. (In vanilla they are red, but this
would conflict with balrogs; orange is a more "aggressive" color
anyway.)