Fix: nemesis didn't deliver dying message
I think this is a bug introduced in the vanilla merge. In vanilla, it
checks for a polymorphed quest leader and directly sets the
killed_leader status to true. In xNetHack, we also check this for a
polymorphed nemesis. But this happens before nemdead() is called, which
checks whether killed_nemesis is true and if so, does nothing else. This
means that the qt_pager for their final speech wasn't getting called.

Call nemdead() (and leaddead(); vanilla probably has a latent bug that
isn't exposed only because leaddead() has no side effects) instead of
directly setting the quest status.