Fix: double pager when killing Surtur on Valk locate level
This would result in Surtur delivering his death speech, then getting a
second pager immediately about the bridge's restoration. Also, the text
of that pager implies you are returning to the level.

Fix this by splitting it into two pager messages: the original one
intended for returning to the level, and the new one intended for being
present to witness the bridge restoration, which is in present tense and
avoids being a second pager by virtue of being one line of text.

We still don't have a great way to detect if you can actually see the
bridge. If you manage to kill Surtur out of sight of the bridge the
message doesn't make any sense. (Also if you're blind - but this is
already a rampant problem in most of the quest text.)