Consistently guard against monsters hiding under trice corpses
There was some code in mon.c in one of the places monsters could decide
to hide which prevented them from doing it if the only thing they could
hide under is corpses which would petrify them, which wasn't
consistently applied.

This moves that check into concealed_spot(), and by extension also
allows it to apply to the hero. Ironically, while that might guard
against a couple places where it's incidental or automatic, I carved out
an exception in the case where the hero is *deliberately* hiding under
cockatrice corpses with #monster, since that's clearly intended
behavior. That will still stone them.