Fix: monster triggering doortrap produced "It crashes on top of it!"
... when a monster triggered a hingeless door trap from the opposite
side from the hero. The point wasn't unblocked, so the hero could still
not "see" the monster and it rendered as "It". I also tweaked it to
hurtle the monster onto the door's space before printing the message.

Adjusted some other code in doortrapped() while here: there is now a
range check on the alternate message to avoid "You hear a nearby crash"
from all the way across the map, and I changed another dist2() into a
distu().