Extend the crackable glass system to all glass weapons and armor
Merge followup 11. This system added in vanilla works great, but would
have an obvious inconsistency if glass weapons weren't given the same
treatment. This also ends up being fairer for users of glass weapons:
instead of a 1/6 chance of completely breaking each hit, the user is now
shown when they are thoroughly cracked and at risk of shattering.

This doesn't completely address xNetHack code around glass item
breakage; that will be in the next commit.