Fix: wishing for cracked glass objects usually didn't work
This occurred because the wishing code attempted to set oeroded after
initially setting the object's material to its object type's base
material, and is_crackable() is false for most such objects. Thus,
wishing for "cracked glass shield of reflection" would be testing
whether to set oeroded on an item for which that doesn't matter. Moving
the oeroded setting after the material setting fixes this issue.

Initially I was going to wontfix this because I thought it would only
ever affect wizard-mode wishes, but then realized a non-wizmode player
can wish for Mirror Brand and if they want it cracked, it ought to work
since you can wish for other artifacts to generate eroded.