Fix: magic flute and harp could be trivially identified
... by trying to call their object class. Since the basic flute and
harp's object definitions had the "known" bit set, these would produce
"That is a silly thing to call", whereas the magic ones would prompt for
a string.

I fixed this first by re-adding the "flute" and "harp" unknown
descriptions and flipping the known bit off, but also going further by
changing the actual names of the basic flute and harp names to "mundane
flute/harp". It's not my favorite adjective, but I decided it was better
than similar adjectives like "simple", "common", "nonmagic", "basic",
"ordinary", and others. It at least implies that there is another type
of harp that isn't mundane, without feeling that overly descriptive.

Considered but rejected at this time was greg88's proposal to add random
shuffled unidentified descriptions for a bunch of different items
(even if just harps and flutes in this case). I still want to do that
but it would be out of scope right now.