Identify an tool the hero is enchanting when it's unambiguous
Noticed when touching the harp/flute constants in this piece of code
that a tool once successfully enchanted does not auto-identify. In
certain cases (horns and bags), this is fine, because the resulting item
is ambiguous, even though it's unambiguous that it was successfully
enchanted into something. But for the other instruments, it's good
enough quality-of-life to automatically identify it when there's no
other possible candidate.

Another related change: when an instrument is enchanted, its known field
is reset to 0 to avoid showing its charges; it needs to be re-identified
in order to see them. This is driven by the ones that enchant into an
ambiguous object type; it's weird to have a fully identified tooled
horn, enchant it, and then not know the new item's object type but know
that it is (0:1) for charges. Also, a sack turning into a bag of tricks
would immediately be obvious since the other bags aren't chargeable.