Show alignment abuse in attribute display
Since there are now mechanics in the game such as erinyes that use
alignment abuse for in-game effects, there should probably be a way for
the player to check it with enlightenment, and it won't hurt to have it
in the dumplog as well.

I don't think "abuse" is a term that is ever shown to the player so I
avoided using it here, instead using "you have been [adverb] untrue to
your alignment", which is a little clunky. Initially I only implemented
this as a binary "abused or not" message, but decided using adverbs to
show the magnitude of abuse is probably more useful to players worried
about erinyes.