Raise the required level to dip for Excalibur in most cases
The reasoning for this is that Excalibur is balanced around not being
known for a large portion of players, or unknown for a substantial
portion of a player's career until they learn about it. This balance
worked well in the 90s, but can't be expected for xNetHack players in
the 2020s where it's reasonable to expect that most are spoiled.

Vanilla partly addressed this in commit e821274: limiting the easiest
access to Excalibur to Knights, and reducing by 80% the chance for all
other characters. However, it meant that it was still possible for early
game lawful non-Knights to get Excalibur simply by being lucky on their
first few dips. The first part of this commit is raising the XL
threshold to 10 for those other roles.

The other part of this commit is factoring in the relatively new
alignment abuse tracking: knights at XL 5-9 will not be able to get
Excalibur if they have abused their alignment even once. This fits in
with Arthurian knight motifs where the greatest rewards and glory go to
only those of greatest virtue or purity. If a Knight has abused their
alignment, they will still be able to dip for it at XL 10 like others.