Rename the Grand Master to the Chan-Sune Lama
"Grand Master" is a very bland name, in the same vein as "The Dark One".
It doesn't help that it collides with a martial arts skill level name.
The heads of Buddhist monasteries are often titled abbots (which
conflicts with the Monk quest guardian name), but some take the title of
lama with the name of their monastery. I don't know exactly where
"Chan-Sune" comes from but it is OK enough to work with without finding
a new name.

"Lama" is an uncommonly seen role title for Priests, but isn't likely to
cause much confusion because they will never see the Chan-Sune Lama.