Overhaul the Monk quest lower filler levels
Monk quest overhaul part 2. These filler levels are underground, part of
the enemy monastery the hero delves into to find Master Kaen, so the
ordinary room-and-corridor fill could technically work, but I believe
that such levels, which are more bland and boring than Dungeons of Doom
levels, should be universally excised from all quests.

The new filler is a fixed (but flippable) system of interconnected
chambers where the areas between them are barred with various
obstructions. I probably could have been achieved true randomness with a
Lua algorithm to randomly place 6x4 squares of ROOM, draw connections,
and then obstruct those connections, but I just did not have the will to
create and debug such a formula.

I took cues on the level's monster stock from the old level, so it's
fairly sparsely populated (except some of the obstructions can randomly
be sleeping earth elementals, so there could be a few more of them).
Loot is sparse, which is a Monk quest theme, but there are a few things
one might expect to find in an "earth temple" sort of level - scrolls of
earth and lots of random gems.

With 5 random trap doors, there's a risk of falling straight into the
Kaen fight if the player isn't careful. That is intended.