Move wizpuzzle data from unsaved-globals struct to a saved-globals one
The recent vanilla merge included the change to global variables where
ones that are saved are stored in distinct structs from ones that
aren't. The variable storing the state of the Wizard's Tower puzzle,
however, created no merge conflicts and thus remained with the non-saved
variables.

This is not technically a bugfix. The restore and save code directly
read/wrote gw.wizpuzzle, which should mean that restored games still
had the correct state. But it's clearly best practice to put structs
that get saved into sv[a-z] instead of g[a-z].