Master Kaen summons elementals instead of insects and can phase
Another component of the Monk quest overhaul. I had considered giving
Kaen a custom spell list which would involve a new elemental-summoning
monster spell, but decided the default clerical spell list is OK if I
just changed summon insects to summon elementals when he's the one
casting it. Since they are stronger than insects, I considered limiting
the amount of elementals the spell creates, but left the default math
mostly in place (usually around d12 elementals, which seems OK since he
has the full clerical menu to cast from and won't pick this that often).
However, when cast at range, he will only summon d2 elementals, to
prevent a situation where the hero is already surrounded, Kaen can't
reach them, and just keeps filling the level with ever more elementals.

His new phasing ability is intended to avoid cheesing the Kaen fight by
use of a boulder fort. Some variants give Kaen the ability to smash
boulders, but as a master of Earth, I figure it's more on theme for him
to be so in tune with the earth that he can pass through stone.

Collectively, these upgrades buff a quest nemesis who was already widely
considered to be the toughest. I am OK with this; it makes the Monk game
more interesting in terms of players deciding to complete post-Quest
objectives before completing the Quest, and gives the difficulty curve
for Monks a different shape.