Overhaul the Monk quest goal level
Next part of the Monk quest overhaul. This is intended to be the
ultimate sanctum of Master Kaen's monastery in the depths of the earth,
and it is shaped like the Chinese character for earth. I removed the
lava, which isn't as thematic and is most likely an artifact of this
level being a copy of the Priest goal level. The monster lineup is
mostly the same, but the fight is now in a smaller and more constrained
space.

As with other revisions in the Monk quest, the random objects are
removed; the main loot on this level is gems and stones, as befitting a
monastery dedicated to Earth. Master Kaen is also carrying the Eyes of
the Overworld instead of having to pick them up, as is the case for
other nemeses of revised quests.