Revise the Monk quest text and change its completion condition
When I updated much of the Monk quest text years ago in an effort to
have it not all be a blatant copy-paste of the Priest quest, I left more
text untouched than I had thought. This cleans out the rest of it. Some
notes:
- Master Kaen is identified as concerning the Chan-Sune Lama for a while
  now, and has great skill at summoning elementals, helping explain why
  they are so prevalent as quest enemies. He wants the Eyes to improve
  this even further by letting him see into the Planes.
- The locate level is named the Monastery of Shi Zhi Zhu, which roughly
  translates to "lords of stone". Quest text now references the locate
  level and the levels before it as such, since the following levels are
  effectively part of its underground complex.
- Master Kaen's maledictions mostly still reflected Nalzok-like demonic
  behavior; they are now befitting a master monk, his earth affinity,
  and his elementals.
- Touching the Eyes of the Overworld now gives you a vision related to
  the quest for the Amulet of Yendor. I felt this was appropriate since
  they do let you see the Astral Plane. In case there is ever anyone
  playing Monk unspoiled, this might also help clue them towards what to
  do.

The condition for completing the Monk quest is now to kill Master Kaen.
Accordingly, there is now an arti_but_not_neme piece of text to catch
the condition where you steal the Eyes and return to the leader while
leaving Master Kaen alive.

I did ultimately have higher hopes of making the story more varied than
"go kill the nemesis", since that's the same as the Valkyrie quest
story and not much different from "go get the artifact", but did not
have an idea for where to take it that I thought was worth implementing
(some potential ideas would have involved breaking many assumptions in
the code about the quest structure and would have required heavy code
changes).