Make Monk home, upper filler, and locate level noflipx
When I initially revised these levels to make them different from the
Priest quest levels, flipping didn't exist. When I merged in flipping
from upstream, I didn't bother updating them because I intended to do it
when updating the Monk quest, and that time is now.

Note that the lower filler does not need to be noflipx because it's an
underground floor and not a rightward trek across the outdoors like most
noflipx quest levels are.