Fix: stray staticfn declarations of not-static-in-xnethack functions
During the recent vanilla merge, I tried to catch these in my own Linux
build with the new USE_NONOSTATICFN. But this evidently didn't work,
and it caused a number of issues I only noticed when I paid attention to
the clang-based MacOS build erroring out on them (as it should).

Now that I fixed USE_NONOSTATICFN in the hints file, staticfn is
correctly getting defined to "static" and gcc is detecting misuse of it
and detecting unused functions. This should fix all the remaining
misuses.