Rewrite the Wizard's Tower teleportation maze
This a complete overhaul of the level, and was not that fun to do, but I
felt I had to do it because it is by far the xNetHack 8.0 feature that
has garnered the most negative feedback. The most common complaints are
that it's too difficult, with a very long, single solution path that is
punishing: if you deviate from it even once, you usually have to find
your way back to the start and try again until through trial and error
you hit the correct 7-13 rooms in sequence.

The major changes I've made are:

1. Reduce the amount of rooms from 27 to 20. This both reduces the
   potential complexity of the maze and allows the rooms to have a
   little more space in them to place things in (more on this below).
2. Reduce the teleporters from 80 one-way teleporters, to 64 two-way
   teleporters (meaning there are only 32 non-directional connections
   between rooms).
3. New algorithm, which takes more lines of code than the old one, but
   that may be because of all the heavy sanity checking I added. This
   algorithm uses a predefined logical graph, the nodes of which various
   physical rooms can be randomly mapped.
4. Lots more stuff in the level to make it less monotonous. Things
   appear in various fixed physical rooms, which you may or may not ever
   see. It also hopefully extends the playing time of this level; when I
   was testing it before stocking anything in the rooms, due to the
   reduced maze complexity, it took only a couple minutes at most of
   wandering around into random teleporters.

There was a change upstream that was little-noticed, but which this
level will probably bring into focus: monsters can track and follow the
player through teleport traps; therefore, escaping through a teleport
trap isn't a safe move as the monster will pop out in pursuit a turn or
two later. I specified all the new monsters as waiting, but they still
seem to start moving around when they hear noise, which means monsters
may randomly jumpscare you by appearing out of nowhere on a teleporter,
and that is honestly pretty on-brand for the Wizard's Tower. It's a place
where the level designer is allowed to be mean to the player.

Thanks to FUZxxl for design discussions about the maze algorithm, and
both them and K2 for suggestions on what to stock some of the rooms
with.